backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
sentence-transformers
faiss-cpu
numpy
openai
